02_api_testing/stellar_burgers_api/requirements.txt
allure-pytest==2.15.2
allure-python-commons==2.15.2
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.1
requests==2.32.5
urllib3==2.6.0
